feat: add checkbox, separate enhanced into nested dropdown story variants
