+++ src/components/reusable/dropdown/dropdownAnchor.scss
@@ -27,7 +27,7 @@
   }
 
   &:hover:not([disabled]) {
-    border-color: var(--kd-color-border-ui-hover);
+    border-color: var(--kd-color-background-ui-hollow-hover);
   }
 
   span {
