+++ src/components/reusable/dropdown/enhancedDropdownOption.scss
@@ -6,18 +6,20 @@
   display: flex;
   align-items: center;
   gap: 12px;
-  padding: 12px 16px;
+  padding: 16px;
   cursor: pointer;
   border: none;
   background: transparent;
   transition: background-color 150ms ease;
   min-height: 74px;
-  background-color: var(
-    --kd-color-background-gradients-interactive-gradient-end
-  );
+  background-color: var(--kd-color-background-forms-default);
+
+  &:active:not([disabled]) {
+    background-color: var(--kd-color-background-accent-subtle);
+  }
 
   &:hover:not([disabled]) {
-    background-color: var(--kd-color-background-accent-subtle);
+    background-color: var(--kd-color-background-ui-hollow-hover);
   }
 
   &[highlighted]:not([disabled]) {
@@ -30,7 +32,11 @@
   }
 
   &[selected]:not([disabled]) {
-    background-color: var(--kd-color-background-ui-hollow-default);
+    background-color: var(--kd-color-background-accent-subtle);
+
+    &:hover {
+      background-color: var(--kd-color-background-ui-hollow-hover);
+    }
   }
 
   &[disabled] {
@@ -90,6 +96,7 @@
 
       ::slotted(kyn-tag) {
         display: inline-flex;
+        pointer-events: none;
       }
     }
   }
@@ -98,7 +105,7 @@
     @include typography.type-ui-03;
     font-weight: 400;
     line-height: 16px;
-    color: var(--kd-color-text-level-secondary);
+    color: var(--kd-color-text-level-primary);
 
     overflow: hidden;
     display: -webkit-box;
@@ -112,7 +119,7 @@
     @include typography.type-ui-03;
     font-weight: 500;
     line-height: 16px;
-    color: var(--kd-color-text-level-primary);
+    color: var(--kd-color-text-level-secondary);
     margin-top: 4px;
 
     overflow: hidden;
@@ -135,7 +142,7 @@
     display: flex;
     width: 16px;
     height: 16px;
-    color: var(--kd-color-support-success);
+    color: var(--kd-color-icon-primary);
   }
 
   .remove-option {
